favorite

+++ resources/views/microposts/microposts.blade.php
@@ -15,7 +15,8 @@
                         <p class="mb-0">{!! nl2br(e($micropost->content)) !!}</p>
                     </div>
                     <div>
-                        @if (Auth::id() == $micropost->user_id)
+                        @include('micropost_favorite.favorite_button')
+                        @if (Auth::id() == $micropost->user_id && Auth::id() == $user->id)
                             {{-- 投稿削除ボタンのフォーム --}}
                             {!! Form::open(['route' => ['microposts.destroy', $micropost->id], 'method' => 'delete']) !!}
                                 {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
